+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- feature/1200197097272268 - Add implode method
 - feature/1200195368152391 - Add toArray, map and filter to IterableHelper, add map and filter to CollectionInterface
 - feature/1200175553157833 - Add iterable helper
 - feature/1200136743167505 - Add properties modifiers to CollectionInterfaceTrait and fix ResetCollectionTrait